refactor : 불필요한 의존성 제거
- UserService에 불필요한 의존성 제거
- 회원입찰관련 서비스를 별도로 구현하여 불필요한 의존성 제거
