+++ src/main/java/org/prgrms/cream/domain/user/dto/UserResponse.java
@@ -1,4 +1,4 @@
-package org.prgrms.cream.domain.user.dto.user;
+package org.prgrms.cream.domain.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
