+++ src/main/java/org/prgrms/cream/domain/user/dto/UserSignUpRequest.java
@@ -1,4 +1,4 @@
-package org.prgrms.cream.domain.user.dto.user;
+package org.prgrms.cream.domain.user.dto;
 
 import lombok.Getter;
 import org.prgrms.cream.domain.user.domain.User;
